refactor: Remove re-export stub files, import directly from shared modules
Delete 4 re-export stub files that existed only as backward compatibility
shims during the manifest-schema consolidation migration:
- src/manifest-editor/schema-helpers.ts
- src/manifest-intellisense/schema-model.ts
- src/manifest-intellisense/xsd-parser.ts
- src/manifest-intellisense/xml-context.ts

All consumers now import directly from src/manifest-schema/.

Copilot-Session: 37e9b215-2ad4-44e8-abc7-3073d1c4edc0
