--- src/manifest-editor/schema-helpers.ts
@@ -1,8 +1,0 @@
-export {
-    getPatternType,
-    buildAttributeFromPatternType,
-    validateValueAgainstType,
-    matchesSchemaPattern,
-    checkSchemaLength,
-    isValidSchemaColor,
-} from '../manifest-schema/schema-helpers';
