--- src/manifest-intellisense/schema-model.ts
@@ -1,14 +1,0 @@
-/**
- * Re-export from shared manifest-schema module.
- * This file exists for backward compatibility during the migration.
- */
-export {
-    SchemaElement,
-    SchemaChildRef,
-    SchemaAttribute,
-    SchemaEnumType,
-    SchemaPatternType,
-    SchemaModel,
-    MANIFEST_NAMESPACES,
-    URI_TO_PREFIX,
-} from '../manifest-schema/schema-model';
