--- src/manifest-intellisense/xsd-parser.ts
@@ -1,5 +1,0 @@
-/**
- * Re-export from shared manifest-schema module.
- * This file exists for backward compatibility during the migration.
- */
-export { loadSchemaModel } from '../manifest-schema/xsd-parser';
